# auth/openapi.json
{"openapi":"3.0.1","info":{"title":"User API","description":"Description","termsOfService":"TOS","contact":{"name":"Baeldung","url":"https://www.baeldung.com","email":"user-apis@baeldung.com"},"license":{"name":"Apache 2.0","url":"https://www.apache.org/licenses/LICENSE-2.0"},"version":"0.0.1"},"servers":[{"description":"Production"}],"paths":{"/register":{"post":{"tags":["auth-controller"],"operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/login":{"post":{"tags":["auth-controller"],"operationId":"token","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/forgotpassword":{"post":{"tags":["auth-controller"],"operationId":"forgotPassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForgotPassword"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/changePassword":{"post":{"tags":["auth-controller"],"operationId":"changePassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePassword"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"Bearer Authentication":[]}]}},"/home":{"get":{"tags":["home-controller"],"operationId":"home","parameters":[{"name":"principal","in":"query","required":true,"schema":{"$ref":"#/components/schemas/UserDetails"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"Bearer Authentication":[]}]}}},"components":{"schemas":{"User":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"},"roles":{"type":"string"},"email":{"type":"string"},"id":{"type":"integer","format":"int64"}}},"LoginRequest":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"ForgotPassword":{"type":"object","properties":{"username":{"type":"string"},"email":{"type":"string"}}},"ChangePassword":{"type":"object","properties":{"currentPassword":{"type":"string"},"newPassword":{"type":"string"}}},"GrantedAuthority":{"type":"object","properties":{"authority":{"type":"string"}}},"UserDetails":{"type":"object","properties":{"enabled":{"type":"boolean"},"authorities":{"type":"array","items":{"$ref":"#/components/schemas/GrantedAuthority"}},"password":{"type":"string"},"username":{"type":"string"},"accountNonLocked":{"type":"boolean"},"accountNonExpired":{"type":"boolean"},"credentialsNonExpired":{"type":"boolean"}}}},"securitySchemes":{"Bearer Authentication":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}